requirements.txt
matplotlib>=3.4.2
Keras~=2.6.0
tensorflow-gpu~=2.6.0
numpy~=1.19.5
tensorflow-datasets~=4.4.0
gdown~=3.13.0
opencv-contrib-python~=4.5.3.56